% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectR.R
\name{kde}
\alias{kde}
\title{Kernel Density Estimation using R}
\usage{
kde(x, h, x_grid)
}
\arguments{
\item{x}{A numeric vector of data points.}

\item{h}{Numeric value representing the bandwidth.}

\item{x_grid}{A numeric vector of points at which to evaluate the density.}
}
\value{
A numeric vector representing the estimated density at each point in x_grid.
}
\description{
Kernel Density Estimation using R
}
