% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lrC}
\alias{lrC}
\title{A Linear regression using Rcpp}
\usage{
lrC(x, y)
}
\arguments{
\item{x}{A numeric vector of independent variable values.}

\item{y}{A numeric vector of dependent variable values.}
}
\value{
A list containing the intercept and slope of the regression line.
}
\description{
A Linear regression using Rcpp
}
\examples{
\dontrun{
y=ax+b
result<-lrC(x,y)
result["intercept"] = b
result["slope"] = a
}
}
